Clean up some duplicate copies of weak.rs.
Both the linux_raw backend and the libc backend had their own copy of
weak.rs; move it out to common code. And, remove tests/process/weak.rs
which is no longer needed.
