+++ Cargo.toml
@@ -48,7 +48,7 @@ libc = { version = "0.2.142", features = ["extra_traits"], optional = true }
 # On all other Unix-family platforms, and under Miri, we always use the libc
 # backend, so enable its dependencies unconditionally.
 [target.'cfg(all(not(windows), any(rustix_use_libc, miri, not(all(target_os = "linux", any(target_arch = "x86", all(target_arch = "x86_64", target_pointer_width = "64"), all(target_endian = "little", any(target_arch = "arm", all(target_arch = "aarch64", target_pointer_width = "64"), target_arch = "powerpc64", target_arch = "riscv64", target_arch = "mips", target_arch = "mips64"))))))))'.dependencies]
-libc_errno = { package = "errno", version = "0.3.0", default-features = false }
+libc_errno = { package = "errno", version = "0.3.1", default-features = false }
 libc = { version = "0.2.142", features = ["extra_traits"] }
 
 # Additional dependencies for Linux with the libc backend:
@@ -78,7 +78,7 @@ default-features = false
 [dev-dependencies]
 tempfile = "3.4.0"
 libc = "0.2.142"
-libc_errno = { package = "errno", version = "0.3.0", default-features = false }
+libc_errno = { package = "errno", version = "0.3.1", default-features = false }
 io-lifetimes = { version = "1.0.10", default-features = false, features = ["close"] }
 # Don't upgrade to serial_test 0.7 for now because it depends on a
 # `parking_lot_core` version which is not compatible with our MSRV of 1.48.
