+++ tests/process/main.rs
@@ -5,10 +5,6 @@
 #![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
 #![cfg_attr(core_c_str, feature(core_c_str))]
 
-#[cfg(not(target_os = "wasi"))]
-#[macro_use]
-mod weak;
-
 mod cpu_set;
 #[cfg(not(target_os = "wasi"))] // WASI doesn't have get[gpu]id.
 mod id;
